+++ ts/packages/knowPro/src/common.ts
@@ -15,8 +15,6 @@ import {
     TextLocation,
     TextRange,
 } from "./interfaces.js";
-import { PropertySearchTerm } from "./interfaces.js";
-import { SearchTermGroup } from "./interfaces.js";
 import { SearchTerm } from "./interfaces.js";
 import {
     IConversationDataWithIndexes,
@@ -98,23 +96,6 @@ export function isInDateRange(outerRange: DateRange, date: Date): boolean {
 
 export function isSearchTermWildcard(searchTerm: SearchTerm): boolean {
     return searchTerm.term.text === "*";
-}
-
-export type TermTypes = SearchTerm | PropertySearchTerm | SearchTermGroup;
-
-export function createAndTermGroup(...terms: TermTypes[]): SearchTermGroup {
-    terms ??= [];
-    return { booleanOp: "and", terms };
-}
-
-export function createOrTermGroup(...terms: TermTypes[]): SearchTermGroup {
-    terms ??= [];
-    return { booleanOp: "or", terms };
-}
-
-export function createOrMaxTermGroup(...terms: TermTypes[]): SearchTermGroup {
-    terms ??= [];
-    return { booleanOp: "or_max", terms };
 }
 
 export async function createConversationFromData(
