+++ ts/packages/knowPro/src/index.ts
@@ -13,6 +13,7 @@ export * from "./propertyIndex.js";
 export * from "./timestampIndex.js";
 export * from "./textLocationIndex.js";
 export * from "./messageIndex.js";
+export * from "./searchCommon.js";
 export * from "./search.js";
 export * from "./serialization.js";
 export * from "./queryCmp.js";
