+++ ts/packages/knowPro/src/searchQueryTranslator.ts
@@ -10,12 +10,8 @@ import {
 import * as querySchema from "./searchQuerySchema.js";
 import { createTypeScriptJsonValidator } from "typechat/ts";
 import { loadSchema } from "typeagent";
-import {
-    createPropertySearchTerm,
-    createSearchTerm,
-    dateRangeFromDateTimeRange,
-    WhenFilter,
-} from "./search.js";
+import { dateRangeFromDateTimeRange, WhenFilter } from "./search.js";
+import { createPropertySearchTerm, createSearchTerm } from "./searchCommon.js";
 import { SearchTermGroup } from "./interfaces.js";
 import { /*isKnownProperty,*/ PropertyNames } from "./propertyIndex.js";
 import { PropertyTermSet } from "./collections.js";
@@ -25,7 +21,7 @@ import {
     createAndTermGroup,
     createOrMaxTermGroup,
     createOrTermGroup,
-} from "./common.js";
+} from "./searchCommon.js";
 
 /*-------------------------------
